+++ libs/co2/data-access/src/lib/co2-forecast/state/create-co2-forecast-interval.spec.ts
@@ -1,4 +1,4 @@
-import { createCo2ForecastDateQuery } from './create-co2-forecast-date-query';
+import { createCo2ForecastDateQuery } from './create-co2-forecast-interval';
 
 describe(createCo2ForecastDateQuery.name, () => {
   it('the start of today is the beginning of the forecast', () => {
