+++ libs/co2/data-access/src/lib/co2-forecast/state/create-co2-forecast-interval.ts
